+++ src/main/java/org/opentestsystem/ap/imrt/iis/IisApplication.java
@@ -1,11 +1,11 @@
-package org.opentestsystem.ap.imrt.iss;
+package org.opentestsystem.ap.imrt.iis;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class IssApplication {
+public class IisApplication {
     public static void main(String[] args) {
-        SpringApplication.run(IssApplication.class, args);
+        SpringApplication.run(IisApplication.class, args);
     }
 }
