+++ src/main/java/org/opentestsystem/ap/imrt/iis/controller/HelloController.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.ap.imrt.iss.controller;
+package org.opentestsystem.ap.imrt.iis.controller;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
