+++ src/test/java/org/opentestsystem/ap/imrt/iis/controller/HelloControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.ap.imrt.iss.controller;
+package org.opentestsystem.ap.imrt.iis.controller;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
